feat: add public brain details modal
