+++ frontend/app/brains-management/library/components/PublicBrainItem/index.ts
@@ -1,0 +1,1 @@
+export * from "./PublicBrainItem";
