+++ frontend/app/brains-management/library/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./PublicBrainItem";
